implementation to align with Jahia GraphQL and OSGi patterns
