+++ graphql-dxm-provider/src/main/java/org/jahia/modules/graphql/provider/dxm/admin/GqlJahiaAdminMutation.java
@@ -21,7 +21,7 @@ import org.apache.commons.lang.StringUtils;
 import org.jahia.modules.graphql.provider.dxm.DataFetchingException;
 import org.jahia.modules.graphql.provider.dxm.osgi.annotations.GraphQLOsgiService;
 import org.jahia.modules.graphql.provider.dxm.osgiconfig.GqlConfigurationMutation;
-import org.jahia.modules.graphql.provider.dxm.service.tags.GqlTagManagerMutation;
+import org.jahia.modules.graphql.provider.dxm.service.tags.graphql.GqlTagManagerMutation;
 import org.jahia.modules.graphql.provider.dxm.util.GqlUtils;
 import org.jahia.services.modulemanager.spi.Config;
 import org.jahia.services.modulemanager.spi.ConfigService;
