--- graphql-dxm-provider/src/main/java/org/jahia/modules/graphql/provider/dxm/service/tags/GqlTagMutationError.java
@@ -1,46 +1,0 @@
-/*
- * Copyright (C) 2002-2022 Jahia Solutions Group SA. All rights reserved.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *   http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.jahia.modules.graphql.provider.dxm.service.tags;
-
-import graphql.annotations.annotationTypes.GraphQLDescription;
-import graphql.annotations.annotationTypes.GraphQLField;
-import graphql.annotations.annotationTypes.GraphQLName;
-import graphql.annotations.annotationTypes.GraphQLNonNull;
-
-@GraphQLName("TagMutationError")
-@GraphQLDescription("An error collected during a tag mutation")
-public class GqlTagMutationError {
-    private final String path;
-    private final String displayableName;
-
-    public GqlTagMutationError(String path, String displayableName) {
-        this.path = path;
-        this.displayableName = displayableName;
-    }
-
-    @GraphQLField
-    @GraphQLNonNull
-    @GraphQLDescription("The impacted node path")
-    public String getPath() {
-        return path;
-    }
-
-    @GraphQLField
-    @GraphQLDescription("The displayable page name when available")
-    public String getDisplayableName() {
-        return displayableName;
-    }
-}
